+++ src/include/mysql_utils.hpp
@@ -64,6 +64,7 @@ public:
 	static string WriteLiteral(const string &identifier);
 	static string EscapeQuotes(const string &text, char quote);
 	static string WriteQuoted(const string &text, char quote);
+	static string TransformConstant(const Value& val);
 };
 
 } // namespace duckdb
